ticketapi/requirements.txt
django
djangorestframework
psycopg2
pytz
httpie
pylint-django